airimapi/requirements.txt
#flask
#flask-cors
fastapi
uvicorn
numpy
pandas
waitress
#azure.storage.blob
PyMuPDF
nltk
python-dotenv
#flask-jwt-extended
langchain
langchain-openai
pymongo
langchain_mongodb
langchain_community
chromadb
#transformers
#torch
pyjwt
python-decouple
